+++ examples/antd-calendar.js
@@ -73,10 +73,13 @@ class Demo extends React.Component {
   constructor(props) {
     super(props);
 
+    this.calendarContainerRef = React.createRef();
+
     this.state = {
       showTime: true,
       showDateInput: true,
       disabled: false,
+      open: false,
       value: props.defaultValue,
     };
   }
@@ -100,6 +103,20 @@ class Demo extends React.Component {
     });
   }
 
+  onOpenChange = (open) => {
+    this.setState({
+      open,
+    });
+  }
+
+  onReadOnlyFocus = () => {
+    this.setState({
+      open: true,
+    });
+  }
+
+  getCalendarContainer = () => this.calendarContainerRef.current;
+
   toggleDisabled = () => {
     this.setState({
       disabled: !this.state.disabled,
@@ -110,7 +127,7 @@ class Demo extends React.Component {
     const state = this.state;
     const calendar = (<Calendar
       locale={cn ? zhCN : enUS}
-      style={{ zIndex: 1000 }}
+      style={{ zIndex: 1001 }}
       dateInputPlaceholder="please input"
       format={getFormat(state.showTime)}
       disabledTime={state.showTime ? disabledTime : null}
@@ -118,6 +135,7 @@ class Demo extends React.Component {
       defaultValue={this.props.defaultCalendarValue}
       showDateInput={state.showDateInput}
       disabledDate={disabledDate}
+      focusablePanel={false}
     />);
     return (<div style={{ width: 400, margin: 20 }}>
       <div style={{ marginBottom: 10 }}>
@@ -161,20 +179,25 @@ class Demo extends React.Component {
           calendar={calendar}
           value={state.value}
           onChange={this.onChange}
+          getCalendarContainer={this.getCalendarContainer}
+          onOpenChange={this.onOpenChange}
+          open={state.open}
+          style={{ zIndex: 1001 }}
         >
           {
             ({ value }) => {
               return (
-                <span tabIndex="0">
-                <input
-                  placeholder="please select"
-                  style={{ width: 250 }}
-                  disabled={state.disabled}
-                  readOnly
-                  tabIndex="-1"
-                  className="ant-calendar-picker-input ant-input"
-                  value={value && value.format(getFormat(state.showTime)) || ''}
-                />
+                <span tabIndex="0" onFocus={this.onReadOnlyFocus}>
+                  <input
+                    placeholder="please select"
+                    style={{ width: 250 }}
+                    disabled={state.disabled}
+                    readOnly
+                    tabIndex="-1"
+                    className="ant-calendar-picker-input ant-input"
+                    value={value && value.format(getFormat(state.showTime)) || ''}
+                  />
+                  <div ref={this.calendarContainerRef} />
                 </span>
               );
             }
@@ -208,7 +231,7 @@ class DemoMultiFormat extends React.Component {
     return (<div style={{ width: 400, margin: 20 }}>
       <div style={{ marginBottom: 10 }}>
         Accepts multiple input formats
-        <br/>
+        <br />
         <small>{multiFormats.join(', ')}</small>
       </div>
       <Calendar
@@ -218,6 +241,7 @@ class DemoMultiFormat extends React.Component {
         format={multiFormats}
         value={state.value}
         onChange={this.onChange}
+        focusablePanel={false}
       />
     </div>);
   }
@@ -256,7 +280,7 @@ ReactDOM.render((<div
         onChange={onStandaloneChange}
         disabledDate={disabledDate}
         onSelect={onStandaloneSelect}
-        renderFooter={() => 'extra footer'}
+        renderFooter={(mode) => (<span>{mode} extra footer</span>)}
       />
     </div>
     <div style={{ float: 'left', width: 300 }}>