+++ src/mixin/CalendarMixin.js
@@ -7,66 +7,58 @@ import { isAllowedDate, getTodayTime } from '../util/index';
 function noop() {
 }
 
-function getNow() {
-  return moment();
-}
-
-function getNowByCurrentStateValue(value) {
+export function getNowByCurrentStateValue(value) {
   let ret;
   if (value) {
     ret = getTodayTime(value);
   } else {
-    ret = getNow();
+    ret = moment();
   }
   return ret;
 }
 
-const CalendarMixin = {
-  propTypes: {
-    value: PropTypes.object,
-    defaultValue: PropTypes.object,
-    onKeyDown: PropTypes.func,
-  },
+export const calendarMixinPropTypes = {
+  value: PropTypes.object,
+  defaultValue: PropTypes.object,
+  onKeyDown: PropTypes.func,
+};
 
-  getDefaultProps() {
-    return {
-      onKeyDown: noop,
-    };
-  },
+export const calendarMixinDefaultProps = {
+  onKeyDown: noop,
+};
 
-  getInitialState() {
-    const props = this.props;
-    const value = props.value || props.defaultValue || getNow();
-    return {
-      value,
-      selectedValue: props.selectedValue || props.defaultSelectedValue,
-    };
-  },
+export const calendarMixinWrapper = ComposeComponent => class extends ComposeComponent {
+  static displayName = 'CalendarMixinWrapper';
+  static defaultProps = ComposeComponent.defaultProps;
 
-  componentWillReceiveProps(nextProps) {
-    let { value } = nextProps;
-    const { selectedValue } = nextProps;
+  static getDerivedStateFromProps(nextProps, prevState) {
+    // Use origin function if provided
+    if (ComposeComponent.getDerivedStateFromProps) {
+      return ComposeComponent.getDerivedStateFromProps(nextProps, prevState);
+    }
+
+    const { value, selectedValue } = nextProps;
+    const newState = {};
+
     if ('value' in nextProps) {
-      value = value || nextProps.defaultValue || getNowByCurrentStateValue(this.state.value);
-      this.setState({
-        value,
-      });
+      newState.value =
+        value || nextProps.defaultValue || getNowByCurrentStateValue(prevState.value);
     }
     if ('selectedValue' in nextProps) {
-      this.setState({
-        selectedValue,
-      });
+      newState.selectedValue = selectedValue;
     }
-  },
 
-  onSelect(value, cause) {
+    return newState;
+  }
+
+  onSelect = (value, cause) => {
     if (value) {
       this.setValue(value);
     }
     this.setSelectedValue(value, cause);
-  },
+  }
 
-  renderRoot(newProps) {
+  renderRoot = (newProps) => {
     const props = this.props;
     const prefixCls = props.prefixCls;
 
@@ -84,24 +76,27 @@ const CalendarMixin = {
         style={this.props.style}
         tabIndex="0"
         onKeyDown={this.onKeyDown}
+        onBlur={this.onBlur}
       >
         {newProps.children}
       </div>
     );
-  },
+  }
 
-  setSelectedValue(selectedValue, cause) {
+  setSelectedValue = (selectedValue, cause) => {
     // if (this.isAllowedDate(selectedValue)) {
     if (!('selectedValue' in this.props)) {
       this.setState({
         selectedValue,
       });
     }
-    this.props.onSelect(selectedValue, cause);
+    if (this.props.onSelect) {
+      this.props.onSelect(selectedValue, cause);
+    }
     // }
-  },
+  }
 
-  setValue(value) {
+  setValue = (value) => {
     const originalValue = this.state.value;
     if (!('value' in this.props)) {
       this.setState({
@@ -110,18 +105,16 @@ const CalendarMixin = {
     }
     if (
       originalValue && value && !originalValue.isSame(value) ||
-        (!originalValue && value) ||
-        (originalValue && !value)
+      (!originalValue && value) ||
+      (originalValue && !value)
     ) {
       this.props.onChange(value);
     }
-  },
+  }
 
-  isAllowedDate(value) {
+  isAllowedDate = (value) => {
     const disabledDate = this.props.disabledDate;
     const disabledTime = this.props.disabledTime;
     return isAllowedDate(value, disabledDate, disabledTime);
-  },
+  }
 };
-
-export default CalendarMixin;